Add files via upload
I am including the research paper you asked for additional points.  I would like to present my paper/presentation at the following conference: Live 360 Orlando it is in November in Orlando and is an industry conference not academic.  Here is the call for presentations website: https://live360events.com/pages/call-for-presentations.aspx. I can't submit the URL  and the paper on Canva so I am going to upload it into the github as well.